% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{approxGradient}
\alias{approxGradient}
\title{Approximate Local Gradient}
\usage{
approxGradient(m)
}
\arguments{
\item{m}{A matrix.}
}
\value{
A list of two matrices \code{r} and \code{c} representing the local
 gradient along the rows and columns of the matrix respectively.
}
\description{
This function computes an approximation of the local 2D gradient
 of a matrix using the Sobel operator.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

