% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc}
\alias{fc}
\title{Foraging Cost Function}
\usage{
fc(l, k, n)
}
\arguments{
\item{l}{The number of larvae in the colony.}

\item{k}{The maximum number of larvae that a colony can have at any given
time.}

\item{n}{A parameter that determines how the cost of foraging scales with the
number of larvae to be fed.}
}
\value{
A vector of the same length as \code{l}.
}
\description{
This function models the foraging cost associated with a given
 number of larvae present in an ant colony.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

