% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lt}
\alias{lt}
\title{Model of Larval Production}
\usage{
lt(t, A, M, P, e)
}
\arguments{
\item{t}{A vector of the times at which to compute the number of larvae.}

\item{A}{The relative amplitude of the larval production cycle.}

\item{M}{The average number of larvae in a colony at any one time.}

\item{P}{The period of the larval production cycle.}

\item{e}{The smoothness of the larval production cycle.}
}
\value{
A vector of the same length as \code{t}.
}
\description{
This function models the number of larvae present in an ant
 colony as function of time.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

