% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tc}
\alias{tc}
\title{Total Foraging Cost}
\usage{
tc(t, A, M, P, e, k, n)
}
\arguments{
\item{t}{A vector of the times at which to compute the number of larvae.}

\item{A}{The relative amplitude of the larval production cycle.}

\item{M}{The average number of larvae in a colony at any one time.}

\item{P}{The period of the larval production cycle.}

\item{e}{The smoothness of the larval production cycle.}

\item{k}{The maximum number of larvae that a colony can have at any given
time.}

\item{n}{A parameter that determines how the cost of foraging scales with the
number of larvae to be fed.}

\item{l}{The number of larvae in the colony.}
}
\value{
A vector of the same length as \code{t}.

A vector of the same length as \code{t}.
}
\description{
This function models the total foraging cost for an ant colony
 as a function of time. It corresponds to the composite of \code{\link{fc}}
 and \code{\link{lt}}.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{fc}}, \code{\link{lt}}
}

